split up giant server file into subroutes, utils, and types
